fix build again

+++ android/app/src/main/kotlin/com/andromapper/map/GeoJsonOverlayManager.kt
@@ -190,7 +190,7 @@ class GeoJsonOverlayManager(
         val paint = AndroidGraphicFactory.INSTANCE.createPaint().apply {
             color = lineColor
         }
-        canvas.drawCircle(size / 2f, size / 2f, size / 2f - 1f, paint)
+        canvas.drawCircle(size / 2, size / 2, (size / 2 - 1).toFloat(), paint)
         return Marker(ll, bitmap, 0, 0)
     }
